@@ -10,25 +10,29 @@ import Admin from "./components/AdminPage/Admin"
 import Movie from './components/Movie/Movie';
 import MoviePage from './components/Moviepage/MoviePage';
 import MovieGrid from './components/MovieGrid/MovieGrid';
+import Search from './components/SearchPage/Search';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { Navbar, Container, Nav, Form, FormControl, Button } from 'react-bootstrap';
 const API_URL = "https://api.themoviedb.org/3/movie/popular?api_key=245b5a23f0b29a2cd2d2fd6c071bad5e";
-//const API_SEARCH = "https://api.themoviedb.org/3/search/movie?api_key=245b5a23f0b29a2cd2d2fd6c071bad5e&query";
+const API_SEARCH = "https://api.themoviedb.org/3/search/movie?api_key=245b5a23f0b29a2cd2d2fd6c071bad5e&query";
+const genre_url = "https://api.themoviedb.org/3/genre/movie/list?api_key=245b5a23f0b29a2cd2d2fd6c071bad5e&language=en-US";
 
 function App() {
 	//const user = localStorage.getItem("token");
 	const [movies, setMovies] = useState([]);
 	const [query, setQuery] = useState('');
-	//const [movieGenre, setMovieGenre] = useState([]);
 
-	// useEffect(() => {
-	// 	fetch(API_URL)
-	// 		.then((res) => res.json())
-	// 		.then(data => {
-	// 			console.log(data);
-	// 			setMovies(data.results);
-	// 		})
-	// }, [])
+	const [movieGenre, setMovieGenre] = useState([]);
+
+
+	useEffect(() => {
+		fetch(API_URL)
+			.then((res) => res.json())
+			.then(data => {
+				console.log(data);
+				setMovies(data.results);
+			})
+	}, [])
 
 
 	const searchMovie = async (e) => {
@@ -39,14 +43,19 @@ function App() {
 			const res = await fetch(url);
 			const data = await res.json();
 			console.log(data);
+
 			console.log("data.results");
 			setMovies(data.results);
 			console.log(movies);
+
+			setMovies(data.results);
+
 		}
 		catch (e) {
 			console.log(e);
 		}
 	}
+
 
 	const searchMovieGenre = async (e) => {
 		e.preventDefault();
@@ -72,24 +81,25 @@ function App() {
 	}
 
 
+
 	const changeHandler = (e) => {
 		setQuery(e.target.value);
 	}
 	return (
 		<div className="App">
-			<Nav />
+			<Nav/>
 
-
-			<BrowserRouter>
+<BrowserRouter>
 				<Routes>
 					{/* user && */}
 					{<Route path="/main" exact element={<Home />} />}
 					<Route path="/signup" exact element={<Signup />} />
 					<Route path="/login" exact element={<Login />} />
 					<Route path="/admin" exact element={<Admin />} />
-					<Route path="/movie" exact component={<MoviePage />} />
+					<Route path="/search" exact element={<Search />} />
+					<Route path="/movie/:id" exact element={<MoviePage />} />
 					{/* <Route path={['/user/:name', '/topics/:name']} component={UsersAndTopics} /> */}
-					<Route path="/" element={<Navigate replace to="/login" />} />
+			{/* <Route path="/" element={<Navigate replace to="/login" />} /> */}
 				</Routes>
 			</BrowserRouter>
 			<Form className="d-flex" onSubmit={searchMovie} autoComplete="off">
@@ -102,6 +112,7 @@ function App() {
 					value={query} onChange={changeHandler}></FormControl>
 				<Button variant="secondary" type="submit">Search</Button>
 			</Form>
+
 			<div>helo</div>
 			<Form className="genres" onSubmit={searchMovieGenre}>
 				<FormControl
@@ -111,6 +122,7 @@ function App() {
 					value={query} onChange={changeHandler}></FormControl>
 				<Button variant="secondary" type="submit">Find Genre</Button>
 			</Form>
+
 
 			<div>
 				{/* <MovieGrid/> */}
